Implement poor man's ngCloak based on app controller
ngCloak is removed by angular when parsing the template, however this
doesn't seem to be enough to keep the content hidden until the
application is completely ready.

We accomplish this by setting `display: none` to the `body` element and
maing it display again when our main controller starts, at which point
we know everything is up and running.