Only append label if this exists in hero-icon
Currently, an empty element was appended to hero-icon template to hold
the label even if this didn't exist, causing some vertical bottom
spacing.

We use Polymer's `dom-if` to conditional show this element. The label
was changed to be passed as an HTML property since it seems to be really
difficult to access the inner HTML of a polymer element programatically.