Rename release/ to etcher-release/
The `bluebird` module contains a directory called "release", which gets
omitted by `electron-packager`.

After some experimentation, there doesn't seems to be a way to make
`electron-packager` ignore a pattern only on the current directory,
therefore we change the name of our output directory to something that
can't possibly collide with something else.