Enable useContentSize BrowserWindow option
From the documentation:

> `useContentSize` Boolean - The `width` and `height` would be used as web
> page’s size, which means the actual window’s size will include window
> frame’s size and be slightly larger. Default is `false`.

The original issue is that when you specify a width/height, the actual
size that you get is slighly smaller, since the OS title bar is included
in the size you provide.

By using the `useContentSize` option, we ensure the `WebView` gets the
intended size, no matter the title bar.

This PR invalidates: https://github.com/resin-io/etcher/pull/244

Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>