Use SUDO_USER environment variable for default file selection path.
Etcher is usually run with sudo as it needs to permission to write
to removable drives. This results to electron opening root user's
file directories by default.

If etcher is run with UID 0 and SUDO_USER is set, with this solution
etcher will open /home/SUDO_USER on linux and /Users/SUDO_USER on
darwin.

This should not have any effect otherwise.