Fix child writer process not working on packaged app
Setting `ELECTRON_RUN_AS_NODE=1` causes the node process to ignore the
default application location ('resources/app.asar') when no argument has
been passed.

This works fine when spawning the child process on develoment mode since
we're pointing to `lib/start.js` manually anyway, but fails when being
packaged since the electron binary tries to run the actual image as a
JavaScript program, failing with weird scary error messages.

As a workaround, we make use of an utility package called
`electron-is-running-in-asar`, which as the name implies, attempts to
detect whether the application is being ran from inside an asar package
or not. If the first case holds, we push the path to the `app.asar`
ourselves.

Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>