fix(GUI): correctly escape single quotes when running inside AppImage
If the application is running from inside an AppImage, the command that
is elevated gets passed as an argument to `sh -c` surrounded by single
quotes.

This means that the usual way of escaping single quotes (with a
backslash), takes no effect:

> Enclosing characters in single quotes preserves the literal value of
> each character within the quotes. A single quote may not occur between
> single quotes, even when preceded by a backslash.
>
> From http://stackoverflow.com/a/1250098/1641422

In this case, the solution is really hacky and obscure: replace single
quotes with `'"'"'`, which ends up being parsed as `'` by the shell.

The following changes have been introduced to accomodate to the above
workaround:

- Rename `utils.escapeWhiteSpacesFromArguments` to
  `utils.escapeArguments`.

- Accept an `willBeSurroundedInSingleQuotes` boolean option in
  `utils.escapeArguments`.

- Make `writer-proxy.js` pass the `willBeSurroundedInSingleQuotes` when
  elevating inside an AppImage.

See: http://stackoverflow.com/a/1250279/1641422
Fixes: https://github.com/resin-io/etcher/issues/738
Change-Type: patch
Changelog-Entry: Fix "Command Failed" error when attempting to flash an image containing a single quote anywhere in the path in GNU/Linux.
Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>